transfert_hubee/requirements.txt
cryptography==3.4.7
python-dotenv==0.18.0
requests==2.25.1
